Move serialization directory out of dags folder
