+++ tests/models/test_serialized_dag.py
@@ -22,8 +22,8 @@
 import unittest
 
 from airflow import example_dags as example_dags_module
-from airflow.dag.serialization.serialized_dag import SerializedDAG
 from airflow.models import DagBag, SerializedDagModel as SDM
+from airflow.serialization.serialized_dag import SerializedDAG
 from airflow.utils import db
